+++ guns-vip-main/src/main/java/cn/stylefeng/guns/meetRegister/model/params/MeetMemberParam.java
@@ -64,7 +64,7 @@ public class MeetMemberParam implements Serializable, BaseValidatingParam {
     /**
      * 缴费记录ID
      */
-    private Integer payId;
+    private Long payId;
 
     @Override
     public String checkParam() {
